refactor(ci): extract the e2e and backend image job groups
The end-to-end job group lived as a private method of the pull-request workflow, which
made it unavailable to any other workflow. It moves to composable builders under
workflows/groups/.

The Management API and Gateway docker images are pulled out of the e2e group into one of
their own. They were only there because e2e was their sole consumer, but the dev-environment
deployment depends on them too: leaving them inside the e2e group means any workflow that
skips e2e also silently loses the images the deployment requires.

Callers contribute the image group before the e2e group, so the emitted job order is
unchanged. Generated configuration is byte-for-byte identical: all 50 fixtures are untouched.
